♻️ Move time utilities from browser-core to @datadog/js-core
Add all time constants and utilities (`dateNow`, `elapsed`, `relativeNow`,
`ONE_SECOND`, `ONE_MINUTE`, etc.) to `packages/js-core/src/time.ts` and
delete the original `packages/browser-core/src/tools/utils/timeUtils.ts`.
